runpod-handler/requirements.txt
# TripoSR RunPod Handler Requirements
# Note: PyTorch should be installed separately with CUDA support

# RunPod SDK
runpod>=1.6.0

# AWS S3 (for large file uploads)
boto3>=1.28.0

# Image processing
Pillow>=10.0.0
rembg[gpu]>=2.0.50

# 3D mesh handling
trimesh>=4.0.0
pyglet<2

# Core dependencies (also in TripoSR requirements.txt)
numpy>=1.24.0
einops>=0.7.0
omegaconf>=2.3.0
transformers>=4.35.0
huggingface-hub>=0.19.0
